#330 started bactesting util functions
